requirements.txt
transformers==4.30.2
torch==2.2.0
faiss-cpu==1.7.3
langchain==0.0.155
cryptography==39.0.1
ipython==8.10.0
#numpy==1.24.3
PyPDF2==3.0.1